Вывод ошибки

+++ app/src/main/java/com/example/filmsapp/domain/responses/OpenFilm.kt
@@ -23,7 +23,7 @@ data class OpenFilm(
         @SerializedName("release_date")
         val date: String,
         @SerializedName("vote_average")
-        val rating: Int
+        val rating: Float
 )
